Update the catagory list and product detail.
