+++ cloudfunction/getProductsList/index.js
@@ -5,6 +5,8 @@ cloud.init({ env: cloud.DYNAMIC_CURRENT_ENV }) // 使用当前云环境
 
 exports.main = async (event, context) => {
   const db = cloud.database();
-  const result = await db.collection('productsList').get();
+  const result = await db.collection('productsList').where({
+    catagoryId: event.catagoryId
+  }).get();
   return result.data;
 }
