+++ components/productDetailComps/productSize/productSize.less
@@ -2,7 +2,7 @@
 .size-layout {
   position: relative;
   height: 100%;
-  padding: 50rpx 40rpx;
+  padding: 80rpx 40rpx 50rpx;
 }
 .selected-layout {
   margin-bottom: 50rpx;
